+++ packages/content-source-maps/src/__tests__/rest.spec.ts
@@ -22,6 +22,10 @@ describe('Content Source Maps with the CPA', () => {
             'en-US': 'English title',
             af: 'Afrikaans title',
           },
+          longText: {
+            'en-US': 'English long text',
+            af: 'Afrikaans long text',
+          },
           richText: {
             'en-US': {
               data: {},
@@ -112,6 +116,12 @@ describe('Content Source Maps with the CPA', () => {
               source: {
                 fieldType: 2,
                 editorInterface: 2,
+              },
+            },
+            'fields/longText': {
+              source: {
+                fieldType: 3,
+                editorInterface: 3,
               },
             },
           },
@@ -211,8 +221,12 @@ describe('Content Source Maps with the CPA', () => {
               widgetId: 'tagEditor',
               widgetNamespace: 'builtin',
             },
+            {
+              widgetId: 'multipleLine',
+              widgetNamespace: 'builtin',
+            },
           ],
-          fieldTypes: ['Symbol', 'RichText', 'Array'],
+          fieldTypes: ['Symbol', 'RichText', 'Array', 'Text'],
         },
       });
 
@@ -251,6 +265,40 @@ describe('Content Source Maps with the CPA', () => {
               widgetNamespace: 'builtin',
             },
             fieldType: 'Symbol',
+          },
+        },
+        '/items/0/fields/longText/en-US': {
+          origin: 'contentful.com',
+          href: 'https://app.contentful.com/spaces/spaceId/environments/master/entries/entryId/?focusedField=longText&focusedLocale=en-US',
+          contentful: {
+            space: 'spaceId',
+            environment: 'master',
+            field: 'longText',
+            locale: 'en-US',
+            entity: 'entryId',
+            entityType: 'Entry',
+            editorInterface: {
+              widgetId: 'multipleLine',
+              widgetNamespace: 'builtin',
+            },
+            fieldType: 'Text',
+          },
+        },
+        '/items/0/fields/longText/af': {
+          origin: 'contentful.com',
+          href: 'https://app.contentful.com/spaces/spaceId/environments/master/entries/entryId/?focusedField=longText&focusedLocale=af',
+          contentful: {
+            space: 'spaceId',
+            environment: 'master',
+            field: 'longText',
+            locale: 'af',
+            entity: 'entryId',
+            entityType: 'Entry',
+            editorInterface: {
+              widgetId: 'multipleLine',
+              widgetNamespace: 'builtin',
+            },
+            fieldType: 'Text',
           },
         },
         '/items/0/fields/richText/en-US/content/0/content/0/value': {
@@ -903,6 +951,134 @@ describe('Content Source Maps with the CPA', () => {
         '/fields/richText/content/0/content/0/value': {
           origin: 'contentful.com',
           href: 'https://app.contentful.com/spaces/spaceId/environments/master/entries/entryId/?focusedField=richText&focusedLocale=en-US',
+          contentful: {
+            space: 'spaceId',
+            environment: 'master',
+            field: 'richText',
+            locale: 'en-US',
+            entity: 'entryId',
+            entityType: 'Entry',
+            editorInterface: {
+              widgetId: 'richTextEditor',
+              widgetNamespace: 'builtin',
+            },
+            fieldType: 'RichText',
+          },
+        },
+      };
+
+      testEncodingDecoding(encodedCPAResponse, mappings);
+    });
+  });
+
+  describe('allows configuring a platform to reduce the payload', () => {
+    const response = createEntry({
+      id: 'entryId',
+      contentType: 'ctId',
+      fields: {
+        richText: {
+          data: {},
+          content: [
+            {
+              data: {},
+              content: [
+                {
+                  data: {},
+                  marks: [],
+                  value: 'Lorem ipsum single locale english',
+                  nodeType: 'text',
+                },
+              ],
+              nodeType: 'paragraph',
+            },
+          ],
+          nodeType: 'document',
+        },
+        title: 'English title',
+      },
+      locale: 'en-US', // sys.locale will be included since we only have one locale in the response
+      contentSourceMapsLookup: {
+        sys: {
+          type: 'ContentSourceMapsLookup',
+        },
+        editorInterfaces: [
+          {
+            widgetId: 'richTextEditor',
+            widgetNamespace: 'builtin',
+          },
+          {
+            widgetId: 'singleLine',
+            widgetNamespace: 'builtin',
+          },
+        ],
+        fieldTypes: ['RichText', 'Symbol'],
+      },
+      contentSourceMaps: {
+        sys: { type: 'ContentSourceMaps' },
+        mappings: {
+          '/fields/richText': {
+            source: {
+              fieldType: 0,
+              editorInterface: 0,
+            },
+          },
+          '/fields/title': {
+            source: {
+              fieldType: 1,
+              editorInterface: 1,
+            },
+          },
+        },
+      },
+    });
+
+    test('setting the platform to vercel removes the contentful information', () => {
+      const encodedCPAResponse = encodeCPAResponse(
+        response,
+        'https://app.contentful.com',
+        'vercel',
+      ) as CPAEntry;
+
+      const mappings = {
+        '/fields/title': {
+          origin: 'contentful.com',
+          href: 'https://app.contentful.com/spaces/spaceId/environments/master/entries/entryId/?focusedField=title&focusedLocale=en-US',
+        },
+        '/fields/richText/content/0/content/0/value': {
+          origin: 'contentful.com',
+          href: 'https://app.contentful.com/spaces/spaceId/environments/master/entries/entryId/?focusedField=richText&focusedLocale=en-US',
+        },
+      };
+
+      testEncodingDecoding(encodedCPAResponse, mappings);
+    });
+
+    test('setting the platform to contentful removes the href information', () => {
+      const encodedCPAResponse = encodeCPAResponse(
+        response,
+        'https://app.contentful.com',
+        'contentful',
+      ) as CPAEntry;
+
+      const mappings = {
+        '/fields/title': {
+          origin: 'contentful.com',
+          contentful: {
+            space: 'spaceId',
+            environment: 'master',
+            field: 'title',
+            locale: 'en-US',
+            entity: 'entryId',
+            entityType: 'Entry',
+            editorInterface: {
+              widgetId: 'singleLine',
+              widgetNamespace: 'builtin',
+            },
+            fieldType: 'Symbol',
+          },
+        },
+        '/fields/richText/content/0/content/0/value': {
+          origin: 'contentful.com',
           contentful: {
             space: 'spaceId',
             environment: 'master',