+++ packages/live-preview-sdk/src/inspectorMode/utils.ts
@@ -1,6 +1,7 @@
 import { decode, type SourceMapMetadata } from '@contentful/content-source-maps';
 import { VERCEL_STEGA_REGEX } from '@vercel/stega';
 
+import { debug } from '../helpers/debug.js';
 import type { InspectorModeOptions } from './inspectorMode.js';
 import {
   InspectorModeAssetAttributes,
@@ -223,6 +224,7 @@ export function getAllTaggedElements({
   for (const { node, text } of stegaNodes) {
     const sourceMap = decode(text);
     if (!sourceMap || !sourceMap.origin.includes('contentful.com')) {
+      debug.warn;
       continue;
     }
 
@@ -257,6 +259,17 @@ export function getAllTaggedElements({
   );
 
   for (const { element, sourceMap } of uniqElementsForTagging) {
+    if (!sourceMap.contentful) {
+      debug.warn(
+        'Element has missing information in their ContentSourceMap, please check if you have restricted the platform for the encoding. (Missing parameter: `contentful`)',
+        {
+          element,
+          sourceMap,
+        },
+      );
+      continue;
+    }
+
     const attributes: InspectorModeSharedAttributes = {
       fieldId: sourceMap.contentful.field,
       locale: sourceMap.contentful.locale,